fix: 1. missing action log for bucket objects ops 2. huawei cloud sync duplicate buckets
