+++ pkg/cloudcommon/db/opslog.go
@@ -214,6 +214,10 @@ const (
 	ACT_GUEST_CREATE_FROM_IMPORT_FAIL    = "guest_create_from_import_fail"
 	ACT_GUEST_PANICKED                   = "guest_panicked"
 	ACT_HOST_MAINTENANCE                 = "host_maintenance"
+
+	ACT_UPLOAD_OBJECT = "upload_obj"
+	ACT_DELETE_OBJECT = "delete_obj"
+	ACT_MKDIR         = "mkdir"
 )
 
 type SOpsLogManager struct {
